microlifx/requirements.txt
pyserial==3.5
lifxlan==1.2.7